fix: fix import data throw 401 error (#1470)
* fix: fix import data throw 401 error

* fix: typo

* fix: typo

+++ packages/admin/src/pages/migration/index.js
@@ -46,11 +46,13 @@ export default function () {
         const tableName = data.tables[i];
         const tableData = data.data[tableName];
 
-        // clean table data
-        await request({
-          url: 'db?table=' + tableName,
-          method: 'DELETE',
-        });
+        // clean table data if not user table
+        if (tableName !== 'Users') {
+          await request({
+            url: 'db?table=' + tableName,
+            method: 'DELETE',
+          });
+        }
 
         if (!idMaps[tableName]) {
           idMaps[tableName] = {};
@@ -60,9 +62,23 @@ export default function () {
         }
 
         for (let j = 0; j < tableData.length; j++) {
+          let exitUserObjectId = false;
+
+          if (tableName === 'Users') {
+            const user = await request('user?email=' + tableData[j].email);
+
+            if (user && user.objectId) {
+              exitUserObjectId = user.objectId;
+            }
+          }
+
+          const method =
+            tableName !== 'Users' || !exitUserObjectId ? 'POST' : 'PUT';
           const resp = await request({
-            url: 'db?table=' + tableName,
-            method: 'POST',
+            url: `db?table=${tableName}${
+              method === 'PUT' ? `&objectId=${exitUserObjectId}` : ''
+            }`,
+            method,
             body: tableData[j],
           });
 
@@ -76,7 +92,7 @@ export default function () {
       }
 
       setImportLoading(['comment data index relationship reconstruction']);
-      const commentData = importData.data.Comment;
+      const commentData = data.data.Comment;
       const willUpdateData = [];
 
       for (let i = 0; i < commentData.length; i++) {
